update reward calculation
